requirements.txt
python-fasthtml
monsterui
feedparser
httpx
pytest
pytest-httpx
playwright
python-dateutil